Update and rename UltraContextIntegration2.py (test).py to UltraContextIntegrationModule.py.py
